+++ docs/ReleaseNotes.md
@@ -51,6 +51,9 @@ line upon naming the release. Refer to previous for appropriate section names.
 - HLSL 202x now supports variadic templates and template function parameter
   packs, pack expansions, and `sizeof...()`
   [#8905](https://github.com/microsoft/DirectXShaderCompiler/issues/8905).
+- HLSL 202x disallows putting cbuffer, tbuffer, or namespace declarations inside
+  a cbuffer or tbuffer
+  [#8484](https://github.com/microsoft/DirectXShaderCompiler/issues/8484).
 
 #### Bug Fixes